Ensure all tasks run to completion
This change raises an exception in all tasks that are active when an
actor is shut down. Previously terminating an actor with uncompleted
calls never sent responses back to actors waiting on those calls.

This also has a side effect of freeing up any JRuby threads that were
being used to execute fibers, as JRuby presently implements fibers with
threads.