Work around issues on Rubinius
On Rubinius I was running into the Thread.current[:task] variable
changing, which may be an issue with Fiber-specific thread local
variables. I'm uncertain at this point, but I thought it might be
related to issues I'm experiencing elsewhere.

Storing the current task in the fiber itself instead of a thread local
works around the problem.